+++ common/matrix_real.cpp
@@ -43,7 +43,7 @@ Matrix_real::Matrix_real() : matrix_base<double>() {
 @param cols_in The number of columns in the stored matrix
 @return Returns with the instance of the class.
 */
-Matrix_real::Matrix_real( double* data_in, size_t rows_in, size_t cols_in) : matrix_base<double>(data_in, rows_in, cols_in) {
+Matrix_real::Matrix_real( double* data_in, int rows_in, int cols_in) : matrix_base<double>(data_in, rows_in, cols_in) {
 
 }
 
@@ -56,7 +56,7 @@ Matrix_real::Matrix_real( double* data_in, size_t rows_in, size_t cols_in) : mat
 @param stride_in The column stride of the matrix array (The array elements in one row are a_0, a_1, ... a_{cols-1}, 0, 0, 0, 0. The number of zeros is stride-cols)
 @return Returns with the instance of the class.
 */
-Matrix_real::Matrix_real( double* data_in, size_t rows_in, size_t cols_in, size_t stride_in) : matrix_base<double>(data_in, rows_in, cols_in, stride_in) {
+Matrix_real::Matrix_real( double* data_in, int rows_in, int cols_in, int stride_in) : matrix_base<double>(data_in, rows_in, cols_in, stride_in) {
 
 }
 
@@ -67,7 +67,7 @@ Matrix_real::Matrix_real( double* data_in, size_t rows_in, size_t cols_in, size_
 @param cols_in The number of columns in the stored matrix
 @return Returns with the instance of the class.
 */
-Matrix_real::Matrix_real( size_t rows_in, size_t cols_in) : matrix_base<double>(rows_in, cols_in) {
+Matrix_real::Matrix_real( int rows_in, int cols_in) : matrix_base<double>(rows_in, cols_in) {
 
 
 }
@@ -80,7 +80,7 @@ Matrix_real::Matrix_real( size_t rows_in, size_t cols_in) : matrix_base<double>(
 @param stride_in The column stride of the matrix array (The array elements in one row are a_0, a_1, ... a_{cols-1}, 0, 0, 0, 0. The number of zeros is stride-cols)
 @return Returns with the instance of the class.
 */
-Matrix_real::Matrix_real( size_t rows_in, size_t cols_in, size_t stride_in) : matrix_base<double>(rows_in, cols_in, stride_in) {
+Matrix_real::Matrix_real( int rows_in, int cols_in, int stride_in) : matrix_base<double>(rows_in, cols_in, stride_in) {
 
 }
 
@@ -126,7 +126,7 @@ Matrix_real::copy() {
 bool
 Matrix_real::isnan() {
 
-    for (size_t idx=0; idx < rows*cols; idx++) {
+    for (int idx=0; idx < rows*cols; idx++) {
         if ( std::isnan(data[idx]) ) {
             return true;
         }
